refactor(input): dissolve default_mode grab-bag into focus and ime features
maintain_input_gates + the webview rect-claim join focus.rs (which defines
the markers they maintain); the ImeCommit->TerminalKeyInput bridge joins
ime.rs. Registrations preserved verbatim (.before(InputPhase::Hover);
observer). Shortcut-applier tests move next to the appliers they test.
