+++ src/input/shortcuts/default_mode.rs
@@ -123,3 +123,255 @@ pub(in crate::input) fn apply_default_type(
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::action::terminal::TerminalSelectionCopy;
+    use crate::input::keyboard::key_effect::KeyEffect;
+    use crate::input::shortcuts::Shortcuts;
+    use crate::surface::OrzmaTerminal;
+    use bevy::ecs::resource::Resource;
+    use bevy::input::keyboard::{Key, KeyCode};
+    use bevy::prelude::{Entity, MinimalPlugins, On, ResMut};
+    use orzma_configs::shortcuts::Modifiers;
+    use orzma_tty_engine::TerminalKey;
+
+    #[derive(Resource, Default)]
+    struct Captured {
+        vi_mode: u32,
+        paste: u32,
+        copy: u32,
+        keys: Vec<TerminalKey>,
+    }
+
+    /// Builds an app running the three Default appliers as bare
+    /// per-message consumers, capturing the events they trigger.
+    fn build_default_dispatch_app(shortcuts: Shortcuts) -> App {
+        let mut app = App::new();
+        app.add_plugins(MinimalPlugins)
+            .add_message::<ShortcutMessage>()
+            .add_message::<ViModeMessage>()
+            .add_message::<TypeMessage>()
+            .init_resource::<Captured>()
+            .insert_resource(shortcuts)
+            .add_systems(
+                Update,
+                (
+                    apply_default_shortcuts,
+                    apply_default_vi_mode.after(apply_default_shortcuts),
+                    apply_default_type
+                        .after(apply_default_shortcuts)
+                        .after(apply_default_vi_mode),
+                ),
+            )
+            .add_observer(|_ev: On<EnterViModeActionEvent>, mut c: ResMut<Captured>| {
+                c.vi_mode += 1;
+            })
+            .add_observer(|_ev: On<PasteAction>, mut c: ResMut<Captured>| {
+                c.paste += 1;
+            })
+            .add_observer(|_ev: On<TerminalSelectionCopy>, mut c: ResMut<Captured>| {
+                c.copy += 1;
+            })
+            .add_observer(|ev: On<TerminalKeyInput>, mut c: ResMut<Captured>| {
+                c.keys.push(ev.key.clone());
+            });
+        app
+    }
+
+    fn default_dispatch_app(shortcuts: Shortcuts) -> (App, Entity) {
+        let mut app = build_default_dispatch_app(shortcuts);
+        let term = app.world_mut().spawn(OrzmaTerminal).id();
+        (app, term)
+    }
+
+    fn meta_mods() -> Modifiers {
+        Modifiers {
+            ctrl: false,
+            shift: false,
+            alt: false,
+            meta: true,
+        }
+    }
+
+    fn dispatch(
+        app: &mut App,
+        effects: Vec<KeyEffect>,
+        focused: Option<Entity>,
+        in_vi_mode: bool,
+        mods: Modifiers,
+    ) {
+        for effect in effects {
+            match effect {
+                KeyEffect::Shortcut { action, via_leader } => {
+                    app.world_mut().write_message(ShortcutMessage {
+                        action,
+                        via_leader,
+                        focused,
+                        in_vi_mode,
+                    });
+                }
+                KeyEffect::ViMode(action) => {
+                    app.world_mut()
+                        .write_message(ViModeMessage { action, focused });
+                }
+                KeyEffect::Type { logical, .. } => {
+                    app.world_mut().write_message(TypeMessage {
+                        logical,
+                        focused,
+                        mods,
+                    });
+                }
+                KeyEffect::WebviewForward { .. } => {}
+            }
+        }
+        app.update();
+    }
+
+    fn type_effect(logical: Key, key_code: KeyCode) -> KeyEffect {
+        KeyEffect::Type { logical, key_code }
+    }
+
+    fn action_effect(action: Shortcut, via_leader: bool) -> KeyEffect {
+        KeyEffect::Shortcut { action, via_leader }
+    }
+
+    #[test]
+    fn plain_key_triggers_terminal_key_input() {
+        let (mut app, term) = default_dispatch_app(Shortcuts::default());
+        dispatch(
+            &mut app,
+            vec![type_effect(Key::Character("a".into()), KeyCode::KeyA)],
+            Some(term),
+            false,
+            Modifiers::default(),
+        );
+        assert_eq!(
+            app.world().resource::<Captured>().keys,
+            vec![TerminalKey::Text("a".into())],
+            "a Type effect must forward to the focused terminal as a TerminalKeyInput"
+        );
+    }
+
+    #[test]
+    fn pane_action_is_noop() {
+        let (mut app, term) = default_dispatch_app(Shortcuts::default());
+        dispatch(
+            &mut app,
+            vec![action_effect(Shortcut::ZoomPane, true)],
+            Some(term),
+            false,
+            Modifiers::default(),
+        );
+        let c = app.world().resource::<Captured>();
+        assert_eq!(
+            (c.vi_mode, c.paste, c.keys.len()),
+            (0, 0, 0),
+            "a Default-mode pane action resolves to a no-op: no event, no typing"
+        );
+    }
+
+    #[test]
+    fn direct_paste_outside_vi_mode_pastes() {
+        let (mut app, term) = default_dispatch_app(Shortcuts::default());
+        dispatch(
+            &mut app,
+            vec![action_effect(Shortcut::Paste, false)],
+            Some(term),
+            false,
+            meta_mods(),
+        );
+        assert_eq!(
+            app.world().resource::<Captured>().paste,
+            1,
+            "a direct paste (via_leader=false) outside vi mode must fire PasteAction"
+        );
+    }
+
+    #[test]
+    fn direct_copy_outside_vi_mode_fires_selection_copy() {
+        let (mut app, term) = default_dispatch_app(Shortcuts::default());
+        dispatch(
+            &mut app,
+            vec![action_effect(Shortcut::Copy, false)],
+            Some(term),
+            false,
+            meta_mods(),
+        );
+        assert_eq!(
+            app.world().resource::<Captured>().copy,
+            1,
+            "a direct copy must fire TerminalSelectionCopy on the focused terminal"
+        );
+    }
+
+    #[test]
+    fn direct_copy_in_vi_mode_also_fires_selection_copy() {
+        let (mut app, term) = default_dispatch_app(Shortcuts::default());
+        dispatch(
+            &mut app,
+            vec![action_effect(Shortcut::Copy, false)],
+            Some(term),
+            true,
+            meta_mods(),
+        );
+        assert_eq!(
+            app.world().resource::<Captured>().copy,
+            1,
+            "copy fires unconditionally: vi mode must not suppress it (unlike paste)"
+        );
+    }
+
+    #[test]
+    fn direct_paste_in_vi_mode_suppressed() {
+        let (mut app, term) = default_dispatch_app(Shortcuts::default());
+        dispatch(
+            &mut app,
+            vec![action_effect(Shortcut::Paste, false)],
+            Some(term),
+            true,
+            meta_mods(),
+        );
+        assert_eq!(
+            app.world().resource::<Captured>().paste,
+            0,
+            "a direct paste in vi mode must be suppressed (via_leader || !in_vi_mode)"
+        );
+    }
+
+    #[test]
+    fn leader_paste_in_vi_mode_pastes() {
+        let (mut app, term) = default_dispatch_app(Shortcuts::default());
+        dispatch(
+            &mut app,
+            vec![action_effect(Shortcut::Paste, true)],
+            Some(term),
+            true,
+            Modifiers::default(),
+        );
+        assert_eq!(
+            app.world().resource::<Captured>().paste,
+            1,
+            "a leader-scoped paste (via_leader=true) must fire even in vi mode"
+        );
+    }
+
+    #[test]
+    fn enter_vi_mode_fires_even_when_already_in_vi_mode() {
+        let (mut app, term) = default_dispatch_app(Shortcuts::default());
+        dispatch(
+            &mut app,
+            vec![action_effect(Shortcut::EnterViMode, false)],
+            Some(term),
+            true,
+            Modifiers::default(),
+        );
+        assert_eq!(
+            app.world().resource::<Captured>().vi_mode,
+            1,
+            "EnterViMode must fire unconditionally in Default mode, even when vi mode is \
+             already active"
+        );
+    }
+}
